remove confidence from property and taxlot
