+++ seed/data_importer/models.py
@@ -875,43 +875,6 @@ class ImportFile(NotDeletableModel, TimeStampedModel):
     def num_cells(self):
         return self.num_rows * self.num_columns
 
-    # TODO: 2/8/17 Verify that this can be removed
-    # @property
-    # def tcm_json(self):
-    #     # JSON used to render the mapping interface.
-    #     tcms = []
-    #     try:
-    #         row_number = 0
-    #         for tcm in self.tablecolumnmappings:
-    #             row_number += 1
-    #             error_message_text = ""
-    #             if tcm.error_message_text:
-    #                 error_message_text = tcm.error_message_text.replace("\n", "<br>")
-    #
-    #             first_rows = ["", "", "", "", ""]
-    #             if tcm.first_five_rows:
-    #                 first_rows = ["%s" % r for r in tcm.first_five_rows]
-    #             tcms.append({
-    #                 'row_number': row_number,
-    #                 'pk': tcm.pk,
-    #                 'destination_model': tcm.destination_model,
-    #                 'destination_field': tcm.destination_field,
-    #                 'order': tcm.order,
-    #                 'ignored': tcm.ignored,
-    #                 'confidence': tcm.confidence,
-    #                 'was_a_human_decision': tcm.was_a_human_decision,
-    #                 'error_message_text': error_message_text,
-    #                 'active': tcm.active,
-    #                 'is_mapped': tcm.is_mapped,
-    #                 'header_row': tcm.first_row,
-    #                 'first_rows': first_rows,
-    #             })
-    #     except:
-    #         from traceback import print_exc
-    #         print_exc()
-    #
-    #     return json.dumps(tcms)
-
     @property
     def tcm_errors_json(self):
         # JSON used to render the mapping interface.
@@ -1234,75 +1197,3 @@ class TableColumnMapping(models.Model):
     def is_mapped(self):
         return self.ignored or (
             self.destination_field is not None and self.destination_model is not None and self.destination_field != "" and self.destination_model != "")
-
-
-class DataCoercionMapping(models.Model):
-    table_column_mapping = models.ForeignKey(TableColumnMapping)
-    source_string = models.TextField()
-    source_type = models.CharField(max_length=50)
-    destination_value = models.CharField(max_length=255, blank=True, null=True)
-    destination_type = models.CharField(max_length=255, blank=True, null=True)
-    is_mapped = models.BooleanField(default=False)
-    confidence = models.FloatField(default=0)
-    was_a_human_decision = models.BooleanField(default=False)
-    valid_destination_value = models.BooleanField(default=False)
-    active = models.BooleanField(default=True)
-
-    def __unicode__(self, *args, **kwargs):
-        return "%s (%s) -> %s (%s)" % (
-            self.source_string, self.source_type, self.destination_value, self.destination_type,)
-
-    def save(self, *args, **kwargs):
-        try:
-            assert self.destination_value is not None
-            field = self.table_column_mapping.destination_django_field
-            field.to_python(self.destination_value)
-            if hasattr(field, "choices") and field.choices != []:
-                assert self.destination_value in [f[0] for f in field.choices] or \
-                    "%s" % self.destination_value in [f[0] for f in field.choices]
-            self.valid_destination_value = True
-        except BaseException:
-            self.valid_destination_value = False
-        self.is_mapped = (
-            self.confidence > 0.6 or self.was_a_human_decision) and self.valid_destination_value
-        super(DataCoercionMapping, self).save(*args, **kwargs)
-
-    @property
-    def source_string_sha(self):
-        if not hasattr(self, "_source_string_sha"):
-            m = hashlib.md5()
-            m.update(self.source_string)
-            self._source_string_sha = m.hexdigest()
-        return self._source_string_sha
-
-
-class ValidationRule(models.Model):
-    table_column_mapping = models.ForeignKey(TableColumnMapping)
-    passes = models.BooleanField(default=False)
-
-
-class RangeValidationRule(ValidationRule):
-    max_value = models.FloatField(blank=True, null=True)
-    min_value = models.FloatField(blank=True, null=True)
-    limit_min = models.BooleanField(default=False)
-    limit_max = models.BooleanField(default=False)
-
-    def __unicode__(self, *args, **kwargs):
-        return "%s<x<%s" % (self.min_value, self.max_value,)
-
-
-class ValidationOutlier(models.Model):
-    rule = models.ForeignKey(ValidationRule)
-    value = models.TextField(blank=True, null=True)
-
-
-class BuildingImportRecord(models.Model):
-    import_record = models.ForeignKey(ImportRecord)
-    building_model_content_type = models.ForeignKey(ContentType, blank=True, null=True)
-    building_pk = models.CharField(max_length=SOURCE_FACILITY_ID_MAX_LEN, blank=True, null=True)
-    building_record = GenericForeignKey('building_model_content_type', 'building_pk')
-    was_in_database = models.BooleanField(default=False)
-    is_missing_from_import = models.BooleanField(default=False)
-
-    def __unicode__(self, *args, **kwargs):
-        return "%s" % self.building_record
